Fix/calculate payouts

+++ feature-staking-impl/src/main/java/jp/co/soramitsu/feature_staking_impl/data/repository/PayoutRepository.kt
@@ -111,7 +111,7 @@ class PayoutRepository(
 
         val nominatorReward = validatorTotalReward * (1 - validatorCommission) * (nominatorStakeInEra / validatorTotalStake)
 
-        return nominatorReward.toInt().toBigInteger()
+        return nominatorReward.toBigDecimal().toBigInteger()
     }
 
     private suspend fun getValidatorHistoricalStats(
